requirements.txt
tensorflow-quantum==0.5.1
gym==0.18.0
pennylane==0.14.1
qiskit==0.32.1
qiskit-machine-learning==0.2.1
torch==1.10.2